renaming app to beta
