+++ android/app/google-services.json.release
@@ -1,0 +1,40 @@
+{
+	"project_info": {
+		"project_number": "694326466848",
+		"firebase_url": "https://nepaltoday.firebaseio.com",
+		"project_id": "nepaltoday",
+		"storage_bucket": "nepaltoday.appspot.com"
+	},
+	"client": [
+		{
+			"client_info": {
+				"mobilesdk_app_id": "1:694326466848:android:a78649736d1c284fc0df45",
+				"android_client_info": {
+					"package_name": "com.siristechnology.nepaltodayapp"
+				}
+			},
+			"oauth_client": [
+				{
+					"client_id": "694326466848-pgl253vpirm9te0ailmolkf3qvvnctnr.apps.googleusercontent.com",
+					"client_type": 3
+				}
+			],
+			"api_key": [
+				{
+					"current_key": "AIzaSyDMBlvCiQQhzSwd8NWRw2mShXIOfWXuPi4"
+				}
+			],
+			"services": {
+				"appinvite_service": {
+					"other_platform_oauth_client": [
+						{
+							"client_id": "694326466848-pgl253vpirm9te0ailmolkf3qvvnctnr.apps.googleusercontent.com",
+							"client_type": 3
+						}
+					]
+				}
+			}
+		}
+	],
+	"configuration_version": "1"
+}
