+++ android/app/src/main/AndroidManifest.xml
@@ -1,4 +1,4 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android" package="com.siristechnology.nepaltodayapp"
+<manifest xmlns:android="http://schemas.android.com/apk/res/android" package="com.siristechnology.nepaltodayapp.beta"
   xmlns:tools="http://schemas.android.com/tools">
 
   <uses-permission android:name="android.permission.INTERNET" />
